# openAPI.yml
...
openapi: 3.1.0
...
components:
  schemas:
...
    LocalDate:
      type: string
      format: date
      examples:
        - 2022-03-10
    LocalDateTime:
      type: string
      format: date-time
      examples:
        - 2022-03-10T12:15:50
    RaccoonUser:
      type: object
      required:
        - email
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        username:
          type: string
        lastfmUsername:
          type: string
        spotifyEnabled:
          type: boolean
        lastNotified:
          $ref: '#/components/schemas/LocalDate'
        createDate:
          $ref: '#/components/schemas/LocalDateTime'
        modifyDate:
          $ref: '#/components/schemas/LocalDateTime'
        lastSpotifyScrape:
          $ref: '#/components/schemas/LocalDateTime'
        lastLastFmScrape:
          $ref: '#/components/schemas/LocalDateTime'
...
    UserSettings:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/RaccoonUser'
        notifyIntervalDays:
          type: integer
          format: int32
        unsubscribed:
          type: boolean
    UserSettingsDto:
      type: object
      properties:
        notifyIntervalDays:
          type: integer
          format: int32
        unsubscribed:
          type: boolean
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://localhost:8181/realms/RaccoonRealm/.well-known/openid-configuration
      description: Authentication
paths:
...
  /me/settings:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSettings'
        required: true
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
      summary: Set User Settings
      tags:
        - User Settings Resource
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettingsDto'
      summary: Get User Settings
      tags:
        - User Settings Resource
...
info:
  title: release-raccoon-app API
  version: 0.4.8-SNAPSHOT